+++ projects/frontend/es-core/src/components/ComponentList.cpp
@@ -1,4 +1,5 @@
 #include <guis/GuiMsgBoxScroll.h>
+#include <Settings.h>
 #include "components/ComponentList.h"
 #include "Util.h"
 #include "Log.h"
@@ -344,7 +345,7 @@ std::vector<HelpPrompt> ComponentList::getHelpPrompts()
 		if(addMovePrompt)
 		  prompts.push_back(HelpPrompt("up/down", _("CHOOSE")));
 	}
-	if(mEntries.at(mCursor).data.help_handler != nullptr){
+	if((mEntries.at(mCursor).data.help_handler != nullptr) && Settings::getInstance()->getInt("HelpPopupTime") != 0){
 		prompts.push_back(HelpPrompt("y", _("HELP")));
 	}